#122 Energy decrease now works perfectly. Bugs fixed.
